Initial modifications to plug JSON composer into extractor; TestExtractor passes on a modified input directory
